Update explainer: Add 'Sec-' prefix to request header
The request header added to requests that are eligible to write to shared storage via response header ought to be a "forbidden header" according to the [Fetch specification](https://fetch.spec.whatwg.org/#forbidden-request-header), i.e. a header that JS cannot add or modify.

We were using `Shared-Storage-Writable` as the name of the request header, which was not a forbidden header and could be modified by JS.

This commit adds the prefix `Sec-` in order to make this header forbidden, updating to using to using `Sec-Shared-Storage-Writable` as the header
name.