Use `sharedStorage` instead of `this.sharedStorage` in worklet
In worklet code,  use `sharedStorage` directly inside of the class methods instead of `this.sharedStorage`, as the latter is incorrect. `this` will refer to the object class rather than the worklet. `globalThis` would be needed instead to refer to the worklet, but `sharedStorage` can simply be referred to directly here.