[Explainer] Support Saved Queries in selectURL
`selectURL()` currently has two per-page-load budgets that restrict that number of calls made to `selectURL()` on each page-load. 
We propose allowing queries to be saved and reused on a per-page basis, where the per-page-load budgets are charged the first time a saved query is run but not for subsequent runs of the saved query during the same page-load. This will be accomplished with a `savedQuery` parameter in the options for `selectURL()` that will name the query.