[Explainer] Mention `resolveToConfig` in list of supported options
`resolveToConfig` was already described as a supported option for `selectURL()` in the explainer, but we omitted to list in the bullet list of supported options. We correct this previous omission. We also slightly clarify the list to make it apparent which options are supported for `run()` and which for `selectURL()`.