[Explainer] Support custom `dataOrigin` in createWorklet
We have developers currently testing the Shared Storage API who have indicated they want to be able to run their worklet on a different origin from the origin that wrote the data. In particular, one of these two actions may be served by, e.g., a CDN, and the other by another domain.  

Third-party cookies allow sharing data at the site level. The current architecture of shared storage, however, only allows the invoking context's origin for `addModule()`/`createWorklet()` to be used as the shared storage data partition origin, or in the case of `createWorklet()`, the worklet script's origin may be manually selected as the data partition origin using the option `dataOrigin: "script-origin"`.

We suggest making it possible for a worklet script to use shared storage data from an arbitrary origin other than the worklet script's origin or the invoking context's origin on an opt-in basis. The data-owning origin would have to opt-in for data use to occur, as data processing could modify the data origin's, send reports, and would affect its shared storage privacy budget. 

While the worklet script will be able to process the data-owning origin's data within a shared storage worklet, the current restrictions will remain in place on output of shared storage data.