[spec] Make bikeshed check references to Private Aggregation API
Currently, all references to the Private Aggregation spec are declared
in the `anchors` block. It's not obvious from the name, but bikeshed
interprets this block as "custom definitions"[1]. Consequently, it has
not been checking our references to Private Aggregation, nor has it
always been generating the correct links[2].

Since the Private Aggregation spec is now in the database[3], I think we
can promote these to real references.

This commit removes the Private Aggregation lines from the `anchors`
block and makes the necessary adjustments to the `link-defaults` block.
With this change, invalid references to the Private Aggregation spec
become compilation errors.

[1]: https://speced.github.io/bikeshed/#custom-dfns
[2]: For instance, clicking "context ID" in step 10 of "obtain the
     pre-specified report parameters" links to a non-existent anchor on
     the Private Aggregation spec.
[3]: https://github.com/tobie/specref/blob/3569d1ecfa298461f3de0586ed2d13ffb000fa29/refs/browser-specs.json#L1179